requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
requests
google-generativeai
python-multipart
openpyxl